requirements.txt
amqp==5.1.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4.4
Deprecated==1.2.13
Django==3.2.15
djangorestframework==3.13.1
Faker==14.1.0
fastapi==0.79.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.30
psycopg2==2.9.3
pydantic==1.9.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.1
